synchrotron.jl Documentation

DiscJetConnections.dn_e — Method

dn_e(γ, n_e0, p, γ_min, γ_max)

Differential electron density at Lotentz factor γ assuming a power law distribution of Lorentz factors from γ_min to γ_max with a power law index p and normalisation n_e0.

The number of electrons with Lorentz factors in the range $\gamma$ to $\gamma + d\gamma$ is given by $dn_e$

\[dn_e = n_{e_0} \gamma^{-p} d\gamma\]

where

\[\gamma_{min} \le \gamma \le \gamma_{max}\]

DiscJetConnections.j_syn — Method

j_syn(ϵ)

Synchrotron emissivity at photon energy ϵ.

For an isotropic electron distribution in a randomly oriented magnetic field, the synchrotron emissivity is (see Dermer et al. (1997) equation 13; we've changed their $H$ to our $B$)

\[j_{syn}(\epsilon, \Omega; x) = \frac{c \sigma_T u_B}{6 \pi \epsilon_B} \left( \frac{\epsilon}{\epsilon_B} \right)^{\tiny{1/2}} \small{n_e} \left[ \left( \frac{\epsilon}{\epsilon_B}\right)^{\tiny{1/2}};x \right]\]

(more documentation here)

DiscJetConnections.S_syn — Method

S_syn()

Should this take parameters (j_syn, z), but these aren't included in the definition?

Synchrotron Flux Density.

\[S_{syn}(\epsilon, \Omega; x) = \frac{D^3 * (1+z) * Vb * j_{syn}}{dL^2}\]

DiscJetConnections.syncPlot — Method

syncPlot()

Example synchotron plot.