Synchrotron_General.jl Documentation

SyncSSCModel.j_nu — Method

j_nu(mps, ν)

General equation of the volume emissivity of synchrotron radiation at frequency ν for parameters mps.

Equation from the book of Longair (1994)

\[j_{\nu} d\nu = -\left( \frac{dE}{dt} \right) N(E) dE \propto \kappa B^{\alpha+1} \nu^{-\alpha} d\nu\]

where, the average energy loss rate for ultra-relativistic electron is equal to

\[-\left( \frac{dE}{dt} \right) = \frac{4}{3} \sigma_T c u_B \gamma^2 \]

N(E) is the electron distribution with power law index p equal to

\[N(E) = \kappa E^{-p}\]

E is the Einstein's mass equation and dE its derivative such that

\[E = \gamma m_e c^2\]

and

\[\alpha = \frac{p-1}{2}\]

synchrotron.jl Documentation

SyncSSCModel.dn_e — Method

dn_e(γ, mps)

Differential electron density at Lorentz factor γ for parameters mps.

The number of electrons with Lorentz factors in the range $\gamma$ to $\gamma + d\gamma$ is given by $dn_e$

\[dn_e = n_{e_0} \gamma^{-p} d\gamma\]

where

\[\gamma_{min} \le \gamma \le \gamma_{max}\]

see also the equation 7.20 in Rybicki and Lightman (1979)

SyncSSCModel.j_syn — Method

j_syn(ϵ, mps)

Synchrotron emissivity at photon energy ϵ for parameters mps.

For an isotropic electron distribution in a randomly oriented magnetic field, the synchrotron emissivity is (see equation 13 of Dermer et al. (1997); we've changed their $H$ to our $B$)

\[j_{syn}(\epsilon, \Omega; x) = \frac{c \sigma_T u_B}{6 \pi \epsilon_B} \left( \frac{\epsilon}{\epsilon_B} \right)^{\tiny{1/2}} \small{n_e} \left[ \left( \frac{\epsilon}{\epsilon_B}\right)^{\tiny{1/2}};x \right]\]

SyncSSCModel.S_syn — Method

S_syn(ϵ, mps)

Synchrotron Flux Density at observed photon energy ϵ.

This impliments equation 3 of Dermer et al. (1997).

\[S_{syn}(\epsilon, \Omega; x) = \frac{D^3 (1+z) V_b j_{syn}(\frac{\epsilon (1+z)}{D}, \Omega; x)}{d_L^2}\]

SyncSSCModel.P_syn — Method

P_syn(ϵ, mps)

Synchrotron Spectral Power Flux at observed photon energy ϵ.

This is equivalent to \nu F(\nu) and presented in equation 23 of Dermer et al. (1997).

\[P_{syn}(\epsilon, \Omega; x) = \epsilon S_{syn}\]

SyncSSCModel.syncSpec — Method

syncSpec(log_ν, mps)

Populate the synchrotron spectrum flux_density with frequency bins given by log_νfor parametersmps``

SyncSSCModel.syncPlot — Method

syncPlot(mps)

Example synchotron plot for parameters mps.

Compton.jl Documentation

SyncSSCModel.j_ssc — Method

j_ssc(ϵ, mps)

Synchrotron Self-Compton (SSC) emissivity at photon energy ϵ for parameters mps.

For an isotropic power-law electron distribution in a randomly oriented magnetic field, the first order SSC emissivity is given by the equation 20 of Dermer et al. (1997)

\[j_{SSC}(\epsilon, \Omega; x) = \frac{c \sigma_T^2 n_{e0}^2 u_B r_b}{9 \pi \epsilon_B} \left( \frac{\epsilon}{\epsilon_B} \right)^{\tiny{-\alpha}} \log{\Sigma_C}\]

SyncSSCModel.P_ssc — Method

P_ssc(ϵ, mps)

Synchrotron Self-Compton (SSC) Spectral Power Flux at observed photon energy ϵ.

This is equivalent to νF(ν) and presented in equation 24 of Dermer et al. (1997).

\[P_{ssc}(\epsilon, \Omega; x) = D^{(3+α)} \frac{c \sigma_T^2 n_{e0}^2 u_B r_b V_b}{9 \pi d_L^2} \left( 1+z \right)^{1 - \alpha} \left( \frac{\epsilon}{\epsilon_B} \right) \ln{\overline{\Sigma_c}}\]

where $\ln{\overline{\Sigma_c}}$ is the transformed Compton-Synchrotron logarithm in equation 25

SyncSSCModel.comptonSpec — Method

comptonSpec(log_ν, mps)

Populate the compton spectrum flux_density with frequency bins given by log_ν for parameters mps

SyncSSCModel.ssc_Plot — Method

ssc_Plot(mps)

Example synchotron self-compton plot for parameters mps.