synchrotron.jl Documentation

DiscJetConnections.dn_e — Method

dn_e(γ, mps)

Differential electron density at Lorentz factor γ for parameters mps.

The number of electrons with Lorentz factors in the range $\gamma$ to $\gamma + d\gamma$ is given by $dn_e$

\[dn_e = n_{e_0} \gamma^{-p} d\gamma\]

where

\[\gamma_{min} \le \gamma \le \gamma_{max}\]

see also the equation 7.20 in Rybicki and Lightman (1979)

DiscJetConnections.j_syn — Method

j_syn(ϵ, mps)

Synchrotron emissivity at photon energy ϵ for parameters mps.

For an isotropic electron distribution in a randomly oriented magnetic field, the synchrotron emissivity is (see equation 13 of Dermer et al. (1997); we've changed their $H$ to our $B$)

\[j_{syn}(\epsilon, \Omega; x) = \frac{c \sigma_T u_B}{6 \pi \epsilon_B} \left( \frac{\epsilon}{\epsilon_B} \right)^{\tiny{1/2}} \small{n_e} \left[ \left( \frac{\epsilon}{\epsilon_B}\right)^{\tiny{1/2}};x \right]\]

DiscJetConnections.S_syn — Method

S_syn(ϵ, mps)

Synchrotron Flux Density at observed photon energy ϵ.

This impliments equation 3 of Dermer et al. (1997).

\[S_{syn}(\epsilon, \Omega; x) = \frac{D^3 (1+z) V_b j_{syn}(\frac{\epsilon (1+z)}{D}, \Omega; x)}{d_L^2}\]

Missing docstring.

Missing docstring for P_syn(ϵ, mps). Check Documenter's build log for details.

DiscJetConnections.syncPlot — Method

syncPlot(mps)

Example synchotron plot for parameters mps.